Revert "Only prompt for recursion if archives are at least 10% of the files."
I do not like this feature. I prefer software that behaves
predictably. This means not defining behavior on the basis of an
arbitrary magic-number algorithm that cannot be readily intuited nor
controlled by the user. The whole point of the interactive UI is to
allow the user to make judgement calls and dictate what happens.

This reverts commit 5c0aff2aef43643108ea83803e8c57b734f11c03.